# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5eb03bec82f82d390e783cd2
  title: Manage_Film_Example
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
  /actor:
    post:
      x-skaffolder-id: 5eb03bed82f82d390e783d00
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5eb03bec82f82d390e783cec
      x-skaffolder-resource: Actor
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Actor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                birthDate:
                  type: integer
                name:
                  type: string
                surname:
                  type: string
    get:
      x-skaffolder-id: 5eb03bed82f82d390e783d08
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5eb03bec82f82d390e783cec
      x-skaffolder-resource: Actor
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Actor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Actor"
  "/actor/{id}":
    delete:
      x-skaffolder-id: 5eb03bed82f82d390e783d09
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5eb03bec82f82d390e783cec
      x-skaffolder-resource: Actor
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Actor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5eb03bed82f82d390e783d03
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5eb03bec82f82d390e783cec
      x-skaffolder-resource: Actor
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Actor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
    post:
      x-skaffolder-id: 5eb03bed82f82d390e783d01
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5eb03bec82f82d390e783cec
      x-skaffolder-resource: Actor
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Actor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Actor"
...
  /user:
    post:
      x-skaffolder-id: 5eb03bec82f82d390e783ce0
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5eb03bec82f82d390e783cdc
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5eb03bec82f82d390e783ce7
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5eb03bec82f82d390e783cdc
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 5eb03bec82f82d390e783ce3
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5eb03bec82f82d390e783cdc
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5eb03bec82f82d390e783ce5
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5eb03bec82f82d390e783cdc
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5eb03bec82f82d390e783ce1
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5eb03bec82f82d390e783cdc
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Actor:
      x-skaffolder-id: 5eb03bec82f82d390e783cec
      x-skaffolder-id-db: 5eb03bec82f82d390e783cd3
      x-skaffolder-id-entity: 5eb03bec82f82d390e783ce8
      x-skaffolder-url: /actor
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        birthDate:
          type: integer
          x-skaffolder-id-attr: 5eb03bec82f82d390e783ceb
          x-skaffolder-type: Date
        name:
          type: string
          x-skaffolder-id-attr: 5eb03bec82f82d390e783ce9
          x-skaffolder-type: String
          x-skaffolder-required: true
        surname:
          type: string
          x-skaffolder-id-attr: 5eb03bec82f82d390e783cea
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
        - name
...
    User:
      x-skaffolder-id: 5eb03bec82f82d390e783cdc
      x-skaffolder-id-db: 5eb03bec82f82d390e783cd3
      x-skaffolder-id-entity: 5eb03bec82f82d390e783cd4
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5eb03bec82f82d390e783cd7
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5eb03bec82f82d390e783cd8
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5eb03bec82f82d390e783cd6
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5eb03bec82f82d390e783cda
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5eb03bec82f82d390e783cd9
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5eb03bec82f82d390e783cd5
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...